lighting: Standardize WAIT value typing
Since millis() gives an unsigned long, use the same type for
the comparator which dictates when to perform an action.

+++ lighting/src/modes.h
@@ -4,13 +4,13 @@ class PixelDispatcher {
     public:
         PixelDispatcher();
         void initialize(int reelsize, int datapin);
-        void runMode(int mode, int wait, int brightness);
+        void runMode(int mode, unsigned long wait, int brightness);
     private:
         Adafruit_NeoPixel strip;
 
         // Delay stuff.
         unsigned long start;
-        bool continueWaiting(int wait);
+        bool continueWaiting(unsigned long wait);
 
         // Rainbow Mode stuff.
         unsigned long rainbow_hue;
